+++ components/auth/NewPassword/NewPasswordContainer.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useState, FC, PropsWithChildren, useEffect } from 'react'
 import { useForm, SubmitHandler } from 'react-hook-form'
 import { useRouter } from 'next/router'
-import { useNewPasswordCreateMutation, useRecoveryCodeCheckMutation } from '@/assets/api/auth/authApi'
+import { useNewPasswordCreateMutation, useRecoveryCodeCheckMutation } from '@/assets/api/auth/authQueryApi'
 import NewPassword from './NewPassword'
 import { IFormInput } from './newPasswordTypes'
 import { CheckRecoveryCode } from './checkRecoveryCode'
