+++ pages/auth/logout/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
 import { getLayout } from '@/components/Layout/Layout'
-import { useLogoutMutation } from '@/assets/api/auth/authApi'
+import { useLogoutMutation } from '@/assets/api/auth/authQueryApi'
 import { authRouts } from '@/components/common/Auth/authRoutes'
 
 const Logout = () => {
@@ -14,7 +14,7 @@ const Logout = () => {
           .unwrap()
           .then(() => {
             router.push(authRouts.login)
-            localStorage.removeItem('accessToken')
+            // localStorage.removeItem('accessToken') //remove from cookie
           })
       } catch (error) {
         console.error('Error occurred while logging out:', error)
