+++ pages/index.tsx
@@ -3,7 +3,7 @@ import { NextPageWithLayout } from './_app'
 import { getLayout } from '@/components/Layout/Layout'
 import Login from './auth/login'
 import { useRouter } from 'next/router'
-import { useMeQuery } from '@/assets/api/auth/authApi'
+import { useMeQuery } from '@/assets/api/auth/authQueryApi'
 import { useEffect } from 'react'
 import { Loading } from '@/components/common/Loaders/Loading'
 import { User, setUser } from '@/assets/api/auth/userSlice'
